requirements.txt
torch>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.35.0
torchxrayvision>=1.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
opencv-python>=4.8.0
scipy>=1.11.0
